fixed errors with internships table

+++ database/migrations/2019_10_10_163300_create_internships_table.php
@@ -16,7 +16,7 @@ class CreateInternshipsTable extends Migration
             $table->string('internship_function');
             $table->text('internship_discription');
             $table->integer('available_spots');
-            $table->bigIncrements('company_id');
+            $table->integer('company_id');
             $table->timestamps();
         });
     }
